pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyapimocker"
version = "0.1.0"
description = "A lightweight, config-driven mock API server for Python teams"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "Jinja2>=3.0.0",
    "Faker>=19.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
]

[project.scripts]
pyapimocker = "pyapimocker.cli:main"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 